requirements.txt
# Flask==2.1.2
# torch==1.12.1
# opencv-python==4.5.3.56  # Changed to an available version
# numpy==1.21.2
# Pillow==8.4.0
# werkzeug==2.1.2
# flask-cors==3.0.10
# ultralytics  # Added for YOLOv8 support

# Core dependencies
Flask==2.1.2
werkzeug==2.1.2
flask-cors==3.0.10

# First install these numeric/scientific packages
numpy>=1.21.5,<1.25.0
Pillow>=8.4.0

# Computer vision
opencv-python>=4.5.3.56

# Deep learning - install separately if needed
# torch>=1.12.1
# torchvision>=0.13.0
ultralytics>=8.0.0  # Will handle its own torch dependencies